capstone_2/pyproject.toml
[project]
name = "traffic-sign-ai"
version = "0.1.0"
requires-python = "==3.11.*"
dependencies = [
    "tensorflow-cpu>=2.16.1",
    "flask",
    "pillow",
    "numpy",
    "pandas",
    "scikit-learn",
    "kagglehub"
]